Ensure always current `stats_start_date` is used
