+++ lib/plausible/stats/query.ex
@@ -121,7 +121,7 @@ defmodule Plausible.Stats.Query do
 
   defp put_period(query, site, %{"period" => "all"}) do
     now = today(site.timezone)
-    start_date = Plausible.Site.local_start_date(site) || now
+    start_date = Plausible.Sites.local_start_date(site) || now
 
     struct!(query,
       period: "all",
